+++ src/main/sanitize.ts
@@ -1,3 +1,4 @@
+import net from 'net';
 import { parsePhoneNumberFromString } from 'libphonenumber-js';
 import type { CountryCode } from 'libphonenumber-js';
 export { validateEmail, validateUrl } from '@shared/validation';
@@ -14,6 +15,54 @@ export function normalizePhone(raw: string, defaultCountry: string = 'US'): stri
     return parsed.formatInternational();
   }
   return null;
+}
+
+/**
+ * Returns true when the URL's hostname resolves to a private/loopback/link-local
+ * address that should be blocked to prevent SSRF.  Uses net.isIP() to classify
+ * addresses rather than regular expressions, which avoids bypass variants such as
+ * `127.1`, `localhost.` (trailing dot), `0.0.0.0`, and long-form IPv6 loopback.
+ *
+ * For hostnames (non-IP strings): only `localhost` (exact, case-insensitive,
+ * after stripping any trailing dots) is blocked — DNS resolution of other names
+ * is left to the OS and is outside the scope of this guard.
+ */
+export function isBlockedHost(urlStr: string): boolean {
+  let hostname: string;
+  try {
+    hostname = new URL(urlStr).hostname;
+  } catch {
+    return true;
+  }
+
+  if (hostname.startsWith('[') && hostname.endsWith(']')) {
+    hostname = hostname.slice(1, -1);
+  }
+
+  const ipVersion = net.isIP(hostname);
+
+  if (ipVersion === 4) {
+    const parts = hostname.split('.').map(Number);
+    const [a, b] = parts;
+    if (a === 127) return true;
+    if (a === 10) return true;
+    if (a === 172 && b >= 16 && b <= 31) return true;
+    if (a === 192 && b === 168) return true;
+    if (a === 169 && b === 254) return true;
+    if (a === 0) return true;
+    return false;
+  }
+
+  if (ipVersion === 6) {
+    const lower = hostname.toLowerCase();
+    if (lower === '::1' || lower === '0:0:0:0:0:0:0:1') return true;
+    const firstGroup = parseInt(lower.split(':')[0] || '0', 16);
+    if ((firstGroup & 0xfe00) === 0xfc00) return true;
+    if ((firstGroup & 0xffc0) === 0xfe80) return true;
+    return false;
+  }
+
+  return hostname.replace(/\.+$/, '').toLowerCase() === 'localhost';
 }
 
 export function detectLinkType(url: string): 'linkedin' | 'x' | 'instagram' | 'facebook' | 'website' {
