+++ src/main/sync/rss-poller.ts
@@ -1,71 +1,11 @@
-import net from 'net';
 import Parser from 'rss-parser';
 import { BrowserWindow, Notification } from 'electron';
 import { v4 as uuidv4 } from 'uuid';
 import { getDatabase, isDatabaseOpen } from '../database';
+import { isBlockedHost } from '../sanitize';
 
 const parser = new Parser();
 
-/**
- * Returns true when the URL's hostname resolves to a private/loopback/link-local
- * address that should be blocked to prevent SSRF.  Uses net.isIP() to classify
- * addresses rather than regular expressions, which avoids bypass variants such as
- * `127.1`, `localhost.` (trailing dot), `0.0.0.0`, and long-form IPv6 loopback.
- *
- * For hostnames (non-IP strings): only `localhost` (exact, case-insensitive,
- * after stripping any trailing dots) is blocked — DNS resolution of other names
- * is left to the OS and is outside the scope of this guard.
- */
-export function isBlockedHost(urlStr: string): boolean {
-  let hostname: string;
-  try {
-    hostname = new URL(urlStr).hostname;
-  } catch {
-    // Invalid URL — block it
-    return true;
-  }
-
-  // Strip surrounding brackets from IPv6 literals (e.g. "[::1]" → "::1")
-  if (hostname.startsWith('[') && hostname.endsWith(']')) {
-    hostname = hostname.slice(1, -1);
-  }
-
-  const ipVersion = net.isIP(hostname);
-
-  if (ipVersion === 4) {
-    const parts = hostname.split('.').map(Number);
-    const [a, b] = parts;
-    // 127.0.0.0/8  — loopback
-    if (a === 127) return true;
-    // 10.0.0.0/8   — RFC-1918 private
-    if (a === 10) return true;
-    // 172.16.0.0/12 — RFC-1918 private
-    if (a === 172 && b >= 16 && b <= 31) return true;
-    // 192.168.0.0/16 — RFC-1918 private
-    if (a === 192 && b === 168) return true;
-    // 169.254.0.0/16 — link-local
-    if (a === 169 && b === 254) return true;
-    // 0.0.0.0/8 — "this" network
-    if (a === 0) return true;
-    return false;
-  }
-
-  if (ipVersion === 6) {
-    const lower = hostname.toLowerCase();
-    // ::1  — loopback (also catches 0:0:0:0:0:0:0:1 after normalization)
-    if (lower === '::1' || lower === '0:0:0:0:0:0:0:1') return true;
-    // fc00::/7 — ULA (first byte 0xfc or 0xfd)
-    const firstGroup = parseInt(lower.split(':')[0] || '0', 16);
-    if ((firstGroup & 0xfe00) === 0xfc00) return true;
-    // fe80::/10 — link-local
-    if ((firstGroup & 0xffc0) === 0xfe80) return true;
-    return false;
-  }
-
-  // Plain hostname — block `localhost` (strip trailing dots first)
-  const stripped = hostname.replace(/\.+$/, '').toLowerCase();
-  return stripped === 'localhost';
-}
 
 function cleanHeadline(raw: string): string {
   return raw
